SampleBinaryClassification/SentimentAnalysis.Data/requirements.txt
astroid==2.2.5
awscli==1.16.255
beautifulsoup4==4.7.1
botocore==1.12.245
certifi==2019.6.16
chardet==3.0.4
Click==7.0
cmake==3.14.4
colorama==0.4.1
cycler==0.10.0
dlib==19.17.0
docutils==0.15.2
face-recognition==1.2.3
face-recognition-models==0.3.0
idna==2.8
imutils==0.5.2
isort==4.3.20
jmespath==0.9.4
joblib==0.13.2
kiwisolver==1.1.0
lazy-object-proxy==1.4.1
matplotlib==3.1.1
mccabe==0.6.1
numpy==1.16.4
oauthlib==3.1.0
opencv-contrib-python==4.1.0.25
pandas==0.24.2
Pillow==6.0.0
pyasn1==0.4.7
pylint==2.3.1
pyparsing==2.4.0
PySocks==1.7.1
python-dateutil==2.8.0
python-dotenv==0.13.0
pytz==2019.1
PyYAML==5.1.2
requests==2.22.0
requests-oauthlib==1.3.0
rsa==3.4.2
s3transfer==0.2.1
scikit-learn==0.21.2
scipy==1.3.0
six==1.12.0
soupsieve==1.9.2
tweepy==3.8.0
twitter==1.18.0
typed-ast==1.4.0
urllib3==1.25.3
virtualenv==16.6.1
wrapt==1.11.2
